+++ inst/shiny/server/server_01_upload.R
@@ -104,6 +104,21 @@ setupSelections <- function() {
         choices = colnames(colData(reactivevalue$se)),
         selected = NULL)
 
+    # umap Analysis
+    updateSelectizeInput(session = session, inputId = 'umap_assay',
+        choices = assayNames((reactivevalue$se)),
+        selected = NULL)
+    updateSelectizeInput(session = session, inputId = 'batch_variable',
+        choices = colnames(colData(reactivevalue$se)),
+        selected = NULL)
+    updateNumericInput(session = session, inputId = 'num_neighbors',
+        value = 15, min = 2,
+        max = dim(reactivevalue$se)[1] - 1)
+    updateNumericInput(session = session, inputId = 'distance',
+        value = 0.1, min = 0.000000000001)
+    updateNumericInput(session = session, inputId = 'spread',
+        value = 1, min = 0.000000000001)
+
     # Variation Analysis
     updateSelectizeInput(session = session, inputId = "variation_assay",
         choices = names(assays(reactivevalue$se)),