+++ inst/shiny/ui/ui_07_pca.R
@@ -1,4 +1,4 @@
-tabPanel("PCA Analysis",
+tabPanel("PCA",
 
          # Application title
          titlePanel("PCA Analysis"),